+++ source/electricity-market/ElectricityMarket.DataApi/Functions/GridAreaTrigger.cs
@@ -15,6 +15,7 @@
 using System.Net;
 using Energinet.DataHub.ElectricityMarket.Application.Commands.GridArea;
 using MediatR;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 
@@ -30,6 +31,7 @@ internal sealed class GridAreaTrigger
     }
 
     [Function(nameof(GetGridAreaAsync))]
+    [Authorize]
     public async Task<HttpResponseData> GetGridAreaAsync(
         [HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "get-grid-area-owner")]
         HttpRequestData req,